refactoring interface adapter files
