--- src/main/java/use_case/timer/SetTimerOutputData.java
@@ -1,4 +1,0 @@
-package use_case.timer;
-
-public class SetTimerOutputData {
-}
